% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{RDatosEcu}
\alias{RDatosEcu}
\title{Author: Diego Guerrero
This function downloads from the RDatosEcu github folder.
Ticket: input a string with all tickets required
Real: transforms nominal variables into real. TRUE, FALSE, or base year. Default is FALSE.
Output: DataFrame file}
\usage{
RDatosEcu(ticket, real = FALSE)
}
\description{
Author: Diego Guerrero
This function downloads from the RDatosEcu github folder.
Ticket: input a string with all tickets required
Real: transforms nominal variables into real. TRUE, FALSE, or base year. Default is FALSE.
Output: DataFrame file
}
\examples{
RDatosEcu("RGDP0000 UNTL1007")
}
