% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{RDatosEcu}
\alias{RDatosEcu}
\title{Author: Diego Guerrero
This function downloads from the RDatosEcu github folder.
Ticket: input a string with all tickets required
Real: transforms nominal variables into real.
Real takes arguments TRUE, FALSE, or a new base year. The default is FALSE.
Output: DataFrame}
\usage{
RDatosEcu(ticket, real = FALSE)
}
\description{
Author: Diego Guerrero
This function downloads from the RDatosEcu github folder.
Ticket: input a string with all tickets required
Real: transforms nominal variables into real.
Real takes arguments TRUE, FALSE, or a new base year. The default is FALSE.
Output: DataFrame
}
\examples{
RDatosEcu("RGDP0000 UNTL1007")
}
