% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDatosEcu.R
\name{RDatosEcu}
\alias{RDatosEcu}
\title{Author: Diego Guerrero
This function downloads from the RDatosEcu github folder
Ticket: input a string with all tickets required
Output: DataFrame file}
\usage{
RDatosEcu(ticket, bulk_file = 1)
}
\description{
Author: Diego Guerrero
This function downloads from the RDatosEcu github folder
Ticket: input a string with all tickets required
Output: DataFrame file
}
\examples{
RDatosEcu("RGDP0000 UNTL1007")
}
